Edit pots controller, db, schema

+++ db/migrate/20250207040459_create_pots.rb
@@ -2,8 +2,8 @@ class CreatePots < ActiveRecord::Migration[7.1]
   def change
     create_table :pots do |t|
       t.string :name
-      t.numeric :total_saved
-      t.numeric :target_amount
+      t.integer :total_saved
+      t.integer :target_amount
 
       t.timestamps
     end
